ENG-22952: widen fs read_file ABI length to u64 to fix wrong-layout dealloc
ex_host_fs_read_file truncated the file length to u32 via `len as u32`
when writing *out_len, and ex_host_free_buffer took a u32 len. The C++
caller (__exactReadFile) copies that many bytes and passes the same
value back to free_buffer, which reconstructs the Box with
slice_from_raw_parts_mut(buf, len) and drops it -- deallocating with a
Layout smaller than the allocation (UB / heap corruption), and silently
short-reading files >= 4 GiB. A file of exactly 4 GiB truncated to 0,
skipping free_buffer's len > 0 guard entirely and leaking the buffer.

Widen the ABI length to u64 on both the Rust side (out_len: *mut u64,
free_buffer len: u64) and every C++ declaration + call site
(hermes_runtime_fs.cc, hermes_runtime_fs_windows.cc, and the stale decl
in hermes_runtime.cc), so the true allocation length round-trips and
free_buffer reconstructs the correct layout. usize -> u64 is lossless on
all supported targets.

Adds a Rust unit test exercising the read_file -> free_buffer round trip.

+++ src/engine/hermes_runtime.cc
@@ -231,9 +231,9 @@ extern "C" int32_t ex_host_fs_read(void* file, uint8_t* buf, uint32_t len);
 extern "C" int32_t ex_host_fs_write(void* file, const uint8_t* buf, uint32_t len);
 extern "C" void ex_host_fs_close(void* file);
 extern "C" uint8_t* ex_host_fs_read_file(const char* path,
-                                         uint32_t* out_len,
+                                         uint64_t* out_len,
                                          int32_t* out_errno);
-extern "C" void ex_host_free_buffer(uint8_t* buf, uint32_t len);
+extern "C" void ex_host_free_buffer(uint8_t* buf, uint64_t len);
 extern "C" char* ex_host_fs_stat(const char* path);
 extern "C" char* ex_host_fs_lstat(const char* path);
 extern "C" char* ex_host_fs_readdir(const char* path);
