+++ src/engine/hermes_runtime_fs.cc
@@ -35,9 +35,9 @@ extern "C" int32_t ex_host_fs_read(void* file, uint8_t* buf, uint32_t len);
 extern "C" int32_t ex_host_fs_write(void* file, const uint8_t* buf, uint32_t len);
 extern "C" void ex_host_fs_close(void* file);
 extern "C" uint8_t* ex_host_fs_read_file(const char* path,
-                                         uint32_t* out_len,
+                                         uint64_t* out_len,
                                          int32_t* out_errno);
-extern "C" void ex_host_free_buffer(uint8_t* buf, uint32_t len);
+extern "C" void ex_host_free_buffer(uint8_t* buf, uint64_t len);
 extern "C" char* ex_host_fs_stat(const char* path);
 extern "C" char* ex_host_fs_lstat(const char* path);
 extern "C" char* ex_host_fs_readdir(const char* path);
@@ -384,7 +384,7 @@ void installFsHostFunctions(ExactHermesRuntime* handle) {
         }
 
         // Single FFI call: read entire file at once
-        uint32_t len = 0;
+        uint64_t len = 0;
         int32_t read_errno = 0;
         uint8_t* buf = ex_host_fs_read_file(path.c_str(), &len, &read_errno);
         if (!buf) {
